Messages do not replace each other
