+++ web-chat/src/Room.tsx
@@ -6,11 +6,12 @@ import {
   ListItemText
 } from '@material-ui/core';
 import React from 'react';
+import { ChatMessage } from 'waku-chat/chat_message';
 import MessageInput from './MessageInput';
 import Send from './Send';
 
 interface Props {
-  lines: string[],
+  lines: ChatMessage[],
 }
 
 interface State {
@@ -58,16 +59,16 @@ export default class Room extends React.Component<Props, State> {
 }
 
 interface LinesProps {
-  messages: string[]
+  messages: ChatMessage[]
 }
 
 const Lines = (props: LinesProps) => {
   const renderedLines = [];
 
-  for (const line of props.messages) {
+  for (const i in props.messages) {
     renderedLines.push(<ListItem>
-      <ListItemText
-        primary={line}
+      <ListItemText key={"chat-message-" + i}
+        primary={printMessage(props.messages[i])}
       />
     </ListItem>);
   }
@@ -78,3 +79,15 @@ const Lines = (props: LinesProps) => {
     </List>
   );
 };
+
+// TODO: Make it a proper component
+function printMessage(chatMsg: ChatMessage) {
+  const timestamp = chatMsg.timestamp.toLocaleString([], {
+    month: 'short',
+    day: 'numeric',
+    hour: 'numeric',
+    minute: '2-digit',
+    hour12: false
+  });
+  return `<${timestamp}> ${chatMsg.nick}: ${chatMsg.message}`;
+}
